change the type passed to display.show()
extmod/modframebuff.c makes use of MP_REGISTER_MODULE() which does
nothing on its own, but that a makemoduledefs.py script scans to
generate the header declaring all the types across all modules:
no header file at all!

Therefore, the types declared for the framebuf are not accessible
from the reset of the C source code base.

I assume it was intended to make use of it by accessing the bytearray
passed to it, which is well defined from C, and that we can use in
our Display.show().

This means the API shown in the example has to be changed.