make fpga.read() take a bytearray instead of returning it
The bytearray.c file does not export any of its functions, only defined
locally, and accessible locally.

This change lets the user create a bytearray, that he would
eventually reuse across calls, and use it to transfer the data
over SPI.

The memory management is left for Micropython to do.